pyproject.toml
[tool.poetry]
name = "binancepythontradealertbot"
version = "0.1.0"
description = ""
authors = ["Replit user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-binance = "^0.7.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
